Refactor: MacheteApiController.cs -> Api/Abstracts
