+++ Machete.Web/Controllers/Api/EmployersController.cs
@@ -4,6 +4,7 @@ using AutoMapper;
 using Machete.Domain;
 using Machete.Service;
 using Machete.Service.DTO;
+using Machete.Web.Controllers.Api.Abstracts;
 using Machete.Web.Helpers.Api;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
