+++ Machete.Web/Controllers/Api/ScheduleRulesController.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using AutoMapper;
 using Machete.Domain;
 using Machete.Service;
+using Machete.Web.Controllers.Api.Abstracts;
 using Machete.Web.Helpers.Api;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
