+++ Machete.Web/Controllers/Api/TransportProvidersController.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using AutoMapper;
 using Machete.Domain;
 using Machete.Service;
+using Machete.Web.Controllers.Api.Abstracts;
 using Machete.Web.Helpers.Api;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
