+++ Machete.Web/Controllers/Api/WorkOrdersController.cs
@@ -2,6 +2,7 @@
 using AutoMapper;
 using Machete.Domain;
 using Machete.Service;
+using Machete.Web.Controllers.Api.Abstracts;
 using Machete.Web.Helpers.Api;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
